requirements.txt
testinfra
pytest-xdist
paramiko
ansible
pyaml
